Web Sales Depot: Quiz Statistic Report 10 fix

+++ WebSalesLibraries/protected/data/tables/tbl_statistic_detail.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS `tbl_statistic_detail`;
 CREATE TABLE IF NOT EXISTS `tbl_statistic_detail` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `id_activity` varchar(32) NOT NULL,
   `tag` varchar(64) NOT NULL,
   `data` varchar(2048) DEFAULT NULL,
